+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add file history
 - [#65] - Add repository to Appveyor and Travis CI (thanks [@papiezak], [@OPNA2608], [@jpcima])
 - [#70] - Add keyboard layout selection for note entry (thanks [@OPNA2608])
+- Add key shortcuts for instrument list
 
 ### Fixed
 - [#69] - Fix corruption in jamming (thanks [@maakmusic], [@ImATrackMan])